+++ src/constants/SalarySettlementPage.ts
@@ -1,3 +1,1 @@
 export const MAX_SALARY: number = 999999999;
-
-export type TMonthSelect = 'prev' | 'next';
